refactor: Добавлен абстрактный компонент. Рефакторинг компонентов с использованием абстрактного компонента. Разделение utils.js на 2 специальных файла.
