+++ src/components/films-section.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import AbstractComponent from './abstract-component.js';
 
 const createFilmsSectionTemplate = () => {
   return (
@@ -8,24 +8,8 @@ const createFilmsSectionTemplate = () => {
 };
 
 
-export default class FilmsSection {
-  constructor() {
-    this._element = null;
-  }
-
+export default class FilmsSection extends AbstractComponent {
   getTemplate() {
     return createFilmsSectionTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
