+++ src/components/header-profile.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import AbstractComponent from './abstract-component.js';
 
 const createHeaderProfileTemplate = () => {
   return (
@@ -10,24 +10,8 @@ const createHeaderProfileTemplate = () => {
 };
 
 
-export default class HeaderProfile {
-  constructor() {
-    this._element = null;
-  }
-
+export default class HeaderProfile extends AbstractComponent {
   getTemplate() {
     return createHeaderProfileTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
