+++ src/components/most-commented-films.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import AbstractComponent from './abstract-component.js';
 
 const createMostCommentedFilmsTemplate = () => {
   return (
@@ -10,24 +10,8 @@ const createMostCommentedFilmsTemplate = () => {
 };
 
 
-export default class MostCommentedFilms {
-  constructor() {
-    this._element = null;
-  }
-
+export default class MostCommentedFilms extends AbstractComponent {
   getTemplate() {
     return createMostCommentedFilmsTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
