+++ src/components/top-rated-films.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import AbstractComponent from './abstract-component.js';
 
 const createTopRatedFilmsTemplate = () => {
   return (
@@ -10,25 +10,13 @@ const createTopRatedFilmsTemplate = () => {
 };
 
 
-export default class TopRatedFilms {
+export default class TopRatedFilms extends AbstractComponent {
   constructor(film) {
+    super();
     this._film = film;
-    this._element = null;
   }
 
   getTemplate() {
     return createTopRatedFilmsTemplate(this._film);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
