+++ src/main.js
@@ -1,7 +1,7 @@
 import {generateFilms} from './mock/films.js';
 import {generateFilters} from './mock/filter.js';
 
-import {renderComponent} from './util.js';
+import {renderComponent} from './utils/render-component.js';
 import NoFilmsComponent from './components/no-films.js';
 import FilmCardComponent from './components/film-card.js';
 import FilmDetailsComponent from './components/film-details.js';
